pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "irminsul"
version = "0.1.0.dev0"
description = "A documentation system for complex codebases. Enforces structural invariants in CI."
readme = "README.md"
requires-python = ">=3.12"
license = "AGPL-3.0-or-later"
license-files = ["LICENSE"]
authors = [{ name = "Irminsul authors" }]
keywords = ["documentation", "docs", "ci", "linter", "mkdocs"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.12",
    "pydantic>=2.7",
    "ruamel.yaml>=0.18",
    "python-frontmatter>=1.1",
    "gitpython>=3.1",
    "pathspec>=0.12",
    "markdown-it-py>=3.0",
    "jinja2>=3.1",
    "litellm>=1.40",
    "httpx>=0.27",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
mkdocs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]

[project.scripts]
irminsul = "irminsul.cli:app"
irm = "irminsul.cli:app"

[project.urls]
Homepage = "https://github.com/huajie-zhong/irminsul"
Repository = "https://github.com/huajie-zhong/irminsul"
Issues = "https://github.com/huajie-zhong/irminsul/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/irminsul"]

[tool.hatch.build.targets.wheel.force-include]
"src/irminsul/init/scaffolds" = "irminsul/init/scaffolds"
"src/irminsul/init/workflows" = "irminsul/init/workflows"

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "B",        # bugbear
    "UP",       # pyupgrade
    "SIM",      # simplify
    "RUF",      # ruff-specific
]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
files = ["src/irminsul"]

[[tool.mypy.overrides]]
module = ["litellm.*", "frontmatter.*", "git.*", "pathspec.*", "markdown_it.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
filterwarnings = ["error"]
